# use vendored version of cupy.pad with added performance optimizations