--- a/python/cucim/src/cucim/skimage/_vendored/__init__.py
+++ b/python/cucim/src/cucim/skimage/_vendored/__init__.py
@@ -3,4 +3,5 @@ This module will hold copies of any upstream CuPy code that is needed, but has
 not yet been merged to CuPy master.
 """
 
+from cucim.skimage._vendored.pad import pad
 from cucim.skimage._vendored.signaltools import *  # noqa