--- a/python/cucim/src/cucim/skimage/measure/_regionprops.py
+++ b/python/cucim/src/cucim/skimage/measure/_regionprops.py
@@ -11,6 +11,7 @@ from scipy.ndimage import find_objects as cpu_find_objects
 
 from . import _moments
 from ._regionprops_utils import euler_number, perimeter, perimeter_crofton
+from cucim.skimage._vendored import pad
 
 __all__ = ['regionprops', 'euler_number', 'perimeter', 'perimeter_crofton']
 
@@ -479,7 +480,7 @@ class RegionProperties:
 
         # TODO: implement marching cubes, etc.
         warn("feret diameter_max currently not implemented on GPU.")
-        identity_convex_hull = cp.pad(
+        identity_convex_hull = pad(
             self.image_convex, 2, mode="constant", constant_values=0
         )
         identity_convex_hull = cp.asnumpy(identity_convex_hull)