--- a/python/cucim/src/cucim/skimage/measure/_regionprops_utils.py
+++ b/python/cucim/src/cucim/skimage/measure/_regionprops_utils.py
@@ -3,6 +3,7 @@ import math
 import cupy as cp
 import cupyx.scipy.ndimage as ndi
 import numpy as np
+from cucim.skimage._vendored import pad
 
 from .._shared.utils import deprecate_kwarg
 
@@ -146,7 +147,7 @@ def euler_number(image, connectivity=None):
 
     # as image can be a label image, transform it to binary
     image = (image > 0).astype(int)
-    image = cp.pad(image, pad_width=1, mode='constant')
+    image = pad(image, pad_width=1, mode='constant')
 
     # check connectivity
     if connectivity is None:
@@ -323,7 +324,7 @@ def perimeter_crofton(image, directions=4):
 
     # as image could be a label image, transform it to binary image
     image = (image > 0).astype(cp.uint8)
-    image = cp.pad(image, pad_width=1, mode="constant")
+    image = pad(image, pad_width=1, mode="constant")
     XF = ndi.convolve(image, cp.array([[0, 0, 0], [0, 1, 4], [0, 2, 8]]),
                       mode='constant', cval=0)