--- a/python/cucim/src/cucim/skimage/measure/block.py
+++ b/python/cucim/src/cucim/skimage/measure/block.py
@@ -1,6 +1,7 @@
 import cupy as cp
 import numpy as np
 
+from .._vendored import pad
 from ..util import view_as_blocks
 
 
@@ -84,8 +85,8 @@ def block_reduce(image, block_size=2, func=cp.sum, cval=0, func_kwargs=None):
             after_width = 0
         pad_width.append((0, after_width))
 
-    image = cp.pad(image, pad_width=pad_width, mode='constant',
-                   constant_values=cval)
+    image = pad(image, pad_width=pad_width, mode='constant',
+                constant_values=cval)
 
     blocked = view_as_blocks(image, block_size)